requirements.txt
joblib==1.1.0
numpy==1.22.2
python-dotenv==0.19.2
scikit_learn==1.0.2
